Allow synchronization to piano rolls
This is a bit of a hack - Dynamic Loncally Normalized Chroma Onsets assume a 12 bin chroma input. This commit makes the size variable depening on an extra argument. This is to allow synchronization against larger data structures such as piano rolls, but it should really rename everything to make that clear.